requirements.txt
PySide6==6.7.2
Pillow==10.4.0
numpy==2.1.1
typing_extensions>=4.9.0
fonttools>=4.53.0

